Rename function to IsInsideExternalXtf
