+++ src/test/data/IsInsideExternalXtf/TestData.xtf
