+++ src/test/data/IsInsideExternalXtf/ch.so.agi.av.hoheitsgrenzen_excerpt.xtf
