+++ src/test/data/IsInsideExternalXtf/swissBOUNDARIES3D_1_3_excerpt.xtf
